perf_events: Update multiplexing code
It turns out the PERF_EVENT_IOC_RESET ioctl resets the count but not
the multiplexing info.

This means that when we fiddle with the events then reset them in
check_scheduability(), we are not really resetting things to zero.

The effect might be small, but since the new multiplex code by
definition is always scheduable, then let's skip the test
if multiplexing.