perf_events: fix some PAPI Error: close of fd = -1 errors
These errors were happening when an event failed to open properly.
(In the cray case, it was perf::PERF_COUNT-HW_BUS_CYCLES not existing).

There was still valid data in the event attribute area, so
when close_pe_events() was called, it tried to close the invalid events.

This fix attempts to only close events that were successfully opened.