add pmu->can_auto_encode() callback
To provide a mechanism for PMU models. to skip
the auto-encoding validation in function
pfmlib_pmu_validate_encoding(). There may be events
for which you cannot simply enumerate all umasks
and encode them one by one. There may be umasks
groups with no defaults. There may be required
modifiers.

In case encoding fails, the validation code now
checks with the PMU layer to verify if the event
does support auto-encoding. That way, the callback
only need to be defined for PMU with known problems.

Signed-off-by: Stephane Eranian <eranian@gmail.com>