simplify perf_util.c loop
There were two separate loops in perf_setup_list_events() that were
iterating pretty much the same way over a string.  To make it a bit
easier to follow the second loop, a for-loop, was rewritten to be as
similar as possible to the earlier while loop.  As a bonus, it keeps
Coverity from complaining about the possible NULL access using the
result of strchr.

Signed-off-by: William Cohen <wcohen@redhat.com>