Fix preset bug
The preset code was only initializing the first element of the
preset code[] array.  Thus any event with more than one subevent
was not terminated at all, and the preset code would use random
garbage as presets.

This exposed another problem; half our code assumed a 0 terminated
code[] array, the rest was looking for PAPI_NULL (-1).  This
standardizes on PAPI_NULL, with comments.

Hopefully this might fix PAPI bug #150.

This is a serious bug and should be included in the next stable release.