Update the memory barriers
It turns out PAPI fails on older 32-bit x86 machines because it tries
to use an SSE rmb() memory barrier.

(Yes, I'm trying to run PAPI on a Pentium II.  Don't ask)

It looks like our memory barriers were copied out of the kernel, which
doesn't quite work because it expects some kernel infrastructure
instead.

This patch uses the definitions used by the "perf" tool instead.

Also dropped the use of the mb() memory barrier on mmap tail write,
as the perf tool itself did a while ago so I'm hoping it's safe to do so
as well.  It makes these definitions a lot simpler.