Remove BITMODE stuff for ppc64 as it is no longer needed
Remove BITMODE stuff for ppc64 as it is no longer needed

libpfm4 is a helper library for use with the perf_events
kernel subsystem, which is only available with relatively
new Linux distributions.  For the ppc64 architecture,
perf_events was first made available in SLES 11 SP1 and
RHEL 6.0.  For both of these distros, the default for
the gcc compiler is to build 64-bit binaries.  This fact,
along with the previous patch in this series having made
LIBDIR behave correctly, means that there's no further
need for the ppc64-specific BITMODE variable to pass
to make.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>