Makefile.inc: Fix library link ordering.
Per Will Cohen
-----------------------------------------------------------
I ran across a problem when trying to build papi with the bundled libpfm and an
earlier incompatible version of libpfm was already installed on the machine.
The make would use the /usr/lib{64}/libpfm.so before trying to use the locally
built version and this would cause problems.

The attached patch changes the order of the linking and uses the local built
libpfm before it tries the installed version.

-Will
-----------------------------------------------------------