Doxygen: Turn off macro expansion for html config.
Doxygen's preprocessor is responsible for things like
--------------------
accumulate and reset hardware events from an event set

< No such EventSet Available

< No error

< No error
...
Definition at line 2619 of file papi.c.
------------------------------

Showing up in the breif description of PAPI_accum in html docs.
This commit enables limiting expansion to a specified set of macros.

What is going on:

int PAPI_api_call() {
..
if(foo)
	return PAPI_OK;
..
}

If doxygen's preprocessor gets ahold of this code,
it becomes
---------
if (foo)
	return 0 /*< No Error */;
---------

Leading to things like;
----------------
< No Error

< No Error
----------------
in the html breifs.