RAPL test code: Add flexibility to the test code.
Per Will Cohen;
------------------
I was reviewing some test results for the papi test and found that the rapl_overflow.c tests makes an assumption that there are exactly two packages. As a result the test will fail on machines with a single package.  The following is a patch to make it a bit more flexible allow 1-n packages in the test.

-Will
-----------------