Fix kernel warning in _papi_hwi_stop_timer()
In _papi_hwi_stop_timer() we were calling
   setitimer( timer, NULL, NULL )
to disable the itimer.

Recent Linux kernels print warnings if you do this; NULL is not
a valid second argument to setitimer() and possibly this wasn't
really working before.

According to the manpage the proper fix is to call setitimer() with
a valid "new_value" field but with the values all 0.  That is what
this patch does.