add ldlat modifier to PEBS-LL events for Intel NHM/WSM/SNB/IVB
ldlat is used to pass the latency threshold
for PEBS Load Latency events on NHM/WSM/SNB/IVB
processors.

ldlat is an integer modifier with valid values
from 3 to 65535. It is only supported by one
event on each processor:
- MEM_INST_RETIRED:LATENCY_ABOVE_THRESHOLD (NHM/WSM)
- MEM_TRANS_RETIRED:LATENCY_ABOVE_THRESHOLD (SNB/IVB)

It must be used with PEBS enabled to trigger the
capture of: instr addr, data addr, load latency,
data source.

Signed-off-by: Stephane Eranian <eranian@gmail.com>