Add component tests' to the install-[all|tests] target.
Thanks to Gary Mohr.
-------------------
This makes a fairly small change to src/Makefile.inc to add logic that adds a
new install-comp_tests target which calls the install target for each component
being built.
This new target is listed as a dependency on the install-tests target so it
will happen when the 'install-all', 'install-tests', or 'install-comp_tests'
targets are used.
A note about this change, I am not real familiar with the auto make and auto
conf tools. This change was enough to make it work for me but if there is
another file that should also be changed for this modification, please help me
out here.

The patch also adds install targets to the Makefiles for all of the components
which have 'tests' directories and updates the README file which talks about
how to create component tests.

Another note, I only compile with a couple of components
(ours, rapl, and example) so if I fat fingered something in one of the other
components Makefiles I would not have noticed.
Please keep me honest and make sure you compile with them all enabled.

Thanks for adding this capability for us.
Gary
---------------------------

Makefile.inc: Add run_tests and friends to install-tests target.

Component test Makefiles' get their install location to mirror what runtests expects.