Coverity fixes:
Coverity pointed out that there was a case where load_user_eent_table()
could leak memory.  The change in the location of the papi_free(foo)
ensures that the allocated memory is freed.

Coverity pointed out one path through the code in
_papi_libpfm4_ntv_code_to_descr() that did not free up memory
allocated in the function.  Added a free on the path in free up
that memory.

Thanks Will Cohen.