cuda/tests/HelloWorld.cu: workaround a segfault.
Report from Gary Hohr
-------------------------------------------------------------------------------
I was running the Cuda test case on a system which did not actually have
any NVIDIA GPU's installed on it (but the cuda software was installed
and papi was built with the cuda component).  I modified the test case to
put an real cuda event in the source (as suggested in the source).  When I
run the test case the cuda component gets disabled in PAPI_library_init
(because detectDevice function can not find any GPU's) which is the
correct behavior. The test case then calls PAPI_event_name_to_code which failed
because the cuda component was disabled.  The test case then created an
event set and called PAPI_add_events with an empty list of events to be
added.  This led to a segfault somewhere inside libpfm4.

The attached patch makes some minor changes to protect against this
problem.  I noticed this test case does not use the PAPI test framework
utilities (test_xxxx functions) so I did not modify the test to use
them.