Cuda component enhancement.
---------------- From Gary's submission---------------------------------
The current packaging of the cuda component in PAPI has a fairly
unfriendly side effect.  When PAPI is built with the cuda component,
then that copy of PAPI can only be used on systems where the cuda
libraries are installed. If it is installed on a system without these
libraries then all PAPI services fail because they have references to
libraries which can not be found.  Even papi_avail which you would think
has nothing to do with cuda reports the error.

This issue significantly complicates the delivery and install of the
PAPI package on large clusters where some of the nodes have NVIDIA GPU's
(and the cuda libraries to talk to them) and other nodes do not have
GPU's (and therefore no software to access them).

I have been working with the help of Phil Mucci to eliminate this
dependency so that a copy of PAPI built with a cuda component could be
installed on all nodes in the cluster and if the node had NVIDIA GPU's
(and libraries available) then the cuda component would get enabled and
could be used.  If the node did not have the hardware or the access
libraries were not available, then the cuda component would just disable
itself at component initialization so it could not be used (but all
other PAPI services would still work).

Phil has provided some gentle prodding and lots of valuable suggestions
to assist this effort.  I now think that I have a working version of
this capability and am ready to share it with the community.
-----------------------------------------------------------------------

Many thanks to Gary Mohr and Phil Mucci for this much needed
functionality.