Address TRAC 174: Let printf do the formatting
https://icl.cs.utk.edu/trac/papi/ticket/174
174: PAPI's debuggin/info output should use %# conversions for octal
and hex
------------------------+--------------------
Reporter:  sbk@…        |      Owner:
   Type:  enhancement  |     Status:  new
Priority:  normal       |  Component:  All
Version:  HEAD         |   Severity:  normal
Keywords:               |
------------------------+--------------------
Email sent to James Ralph:

Seeing your latest change reminded me: Anytime there is a value issued
in hex or octal the "%#" conversion should be used so the value is
always preceded with a "0" for octal or a "0x" for hex. Otherwise when
a value is printed one can not tell the base it is in (one shouldn't
have to rely on internal knowledge of the code or the context to tell).
For variables that are pointers the "%p" conversion can be used (this
will always use an hex syntax).

It would be nice to apply this to all PAPI print statements in their
entirety.