Cuda component: Update library search path
From Gary Mohr:
It turns out that with the changes I gave you the path to the libcuda.so
library is still hard coded to /usr/lib64.  This assumes that the
NVIDIA-Linux package is installed on the system where the build is being
done.  In Bull's case (and probably other users also) this is not always
the case.  To add the flexibility we need, I have added a new configure
argument to the cuda configure script.  The new argument is
"--with-cudrv_dir" and it allows the user to specify where the cuda driver
package (ie: NVIDIA-Linux) to be used for the build can be found.  This new
argument is optional and if not provided a value of "/usr" will be used.
This allows existing configure calls to continue to work like before.