Some more ctest fixes involving disabled components.
We enforce disabled components sometime in the PAPI routines and sometimes
in the components themselves.  A bit confusing.

It is tough with perf_event and perf_event_uncore because we share libpfm4
by both, so the naming library for perf_event_uncore will be active
even if the component is disabled, which can cause some confusing results
if your test code ignores PAPI_ENOCMP error messages and accesses a disabled
component anyway.

This at least fixes our test cases, we might have to revisit this later.