net components: coverity fixes
Reported/patched by Gary Mohr
-----------------------------
The net component has one defect in one of the test cases.  The complaint
is that there is code that can never be executed.  There is a test to see
if event_count == 0 which can never be true at that place in the code.  So
I removed the if statement and its contents.  Here is the patch:
-----------------------------