Infiniband component: switch over to weak linking
Thnaks to Gary Mohr for the patch.
----------------------------------
The infiniband component needs include files and libraries from both the infiniband ibmad and ibumad
packages.  When these packages are installed on a system, both packages normally install their files in the same
place (includes in /usr/include/infiniband and libraries in /usr/lib64).  The current component configure script
allows you to provide a single include path and a single library path which gets used to access files from both
packages.  If these two packages have different install prefixes (or you are trying to build from install images
of each package which are not located under the same directory) then the configure script fails because it can
not find all the files it needs.

These changes modify the configure script to replace the include and library dir's with an ibmad_dir and ibumad_dir
and then uses the correct packages directory when looking for includes and libraries from that package.  This makes
it work like the cuda and nvml components with respect to configuring how to find files from a package the component
depends on.

There are also changes in this patch file to remove an unneeded variable in the dlopen code to resolve some
defects reported by coverity.