ctests: Skip calling into disabled components.
This patch fixes a problem that was causing two test cases to abort when they were
run on a system which has disabled components.  Code was added to check if the component
is disabled and just go to the next component in the list when the check is true.  This
prevents calls to code in components which may abort because the component was unable to
initialize itself correctly.

Thanks to Gary Mohr and Chuck LaCasse from Bull for reporting.