lustre component: fix a free issue.
Reported by Gary Mohr
---------------------
This patch corrects an abort when running the papi utilities (papi_avail, papi_component_avail, ...).
The problem happens when papi is built with the lustre component and is installed on a system where
this component will be enabled (lustre software and file systems are installed).

The abort is caused by the component init function allocating space and putting its pointer on a
linked list.  Then the space the pointer referres to is freed but the linked list still exists.  In
shutdown component function it walks this linked list trying to free the pointers again.  This can
cause either seg faults or aborts in glibc because its free function is passed a bad pointer.

The call to do this free was added by commit a3bda01 (done 02/2012).  The original change was intended
to resolve coverity defects but the change inserted one to many free calls.
---------------------