native_avail: Fix parse_unit_mask code
Reported by Steve Kaufmann
--------------------------
I noticed while developing a new component that the output from
papi_native_avail was incorrectly presented for the component. I believe
this is because the ":::" prefix is not being taken into account, so
the base event name is interpreted as a unit mask and is prepend with a
:
before each legitimate unit mask associated with the event. I think this
is just now happening because mine is the first component that has
unit masks. I have include a fix below. The output of the unit masks by
papi_native_avail now appears correctly for my component.

Thanks, Steve