Clean up option handling in papi_cost
The papi_cost used strstr to seach for the substring that matched the
option.  this is pretty inexact.  Made sure that the options matched
exactly and the option argments for -b and -t were greater than 0.
Also make papi_cost print out the help if there was an option that it
didn't understand.

Signed-off-by: William Cohen <wcohen@redhat.com>