Perf_event_amd_northbridge_test: Use buffer event_name instead of uncore_event
The variable uncore_event is initialized to NULL and is never changed
during execution of the test. PAPI_add_named_event fails and the event set
cannot be started. The correct event name is stored in event_name, replacing
all occurrences of uncore_event with event_name therefore fixes the problem
metioned above.