Fwd: Re: [Ptools-perfapi] ctests/shlib FAILED
Should have sent this to the papi devel list. -Will

-------- Original Message --------
Subject: Re: [Ptools-perfapi] ctests/shlib FAILED
Date: Tue, 09 Jul 2013 23:20:10 -0400
From: William Cohen <wcohen@redhat.com>
To: ptools-perfapi@eecs.utk.edu

On 03/09/2012 03:40 PM, William Cohen wrote:
> I was looking through the test results and found that ctests/shlib FAILED on all the machines I tested on because libm shared library is already linked in. There is no difference in the number of shared libraries before and after the dlopen. The test ctests/shlib fails as a reult of this.
>
> -Will
> _______________________________________________
> Ptools-perfapi mailing list
> Ptools-perfapi@eecs.utk.edu
> http://lists.eecs.utk.edu/mailman/listinfo/ptools-perfapi
>

I did some more investigation of this problem today.  I found that the lmsensor component implicitly pulls in the libm.  As an alternative, I wrote the attached patch that uses setkey() and encrypt() in libcrypt.so instead.  It works on various linux machines, but I do not know whether it is going to work on other OS.

-Will

>From c53c97e1de2d1c7dc0bca64d1906287ff73343c6 Mon Sep 17 00:00:00 2001
From: William Cohen <wcohen@redhat.com>
Date: Tue, 9 Jul 2013 22:37:27 -0400
Subject: [PATCH] Avoid using libm.so for ctests/shlib because of implicit use
 in some components

The lmsensors component can implicitly pull in libm.so into the
executable.  Unfortunately, the ctests/shlib test expects that libm.so
is not loaded and will fail because there is no change in the count of
shared libraries.  The patch uses libcrypt.so library setkey and
encrypt functions to test PAPI_get_shared_lib_info( ) instead of
libm.so library pow function.

Signed-off-by: William Cohen <wcohen@redhat.com>