perf_event/perf_event_uncore/libpfm4 -- rearrange files
Give perf_event and perf_event_uncore copies of papi_libpfm4_events
to work with, as they will have different needs for the code.

Get rid of the perf_event_lib stuff.  It was a hack to begin with
and in the end not much code will be shared.  Maybe we can
re-share things once uncore support is complete.