Attempt to fix a memory leak in fork2 test.
Fork2 does the following:
 PAPI_library_init()
         fork();
         /    \
     parent   child
     wait()     PAPI_shutdown()	-> _papi_hwi_shutdown_global_threads() -> foreach(threadinfo we allocated): _papi_hwi_shutdown_thread()
                PAPI_library_init()

_papi_hwi_shutdown_thread checks who allocated a ThreadInfo entry in the global list,
and will only free it if our thread did the allocation.

When threading is not initialized, we fall back to getpid(), now in the child process,
the one ThreadInfo item on the list was allocated by our parent, so at shutdown time
we don't free this, and thus leak it.

Solution is to add a parameter to _hwi_shutdown_thread to force shutdown even if we
didn't allocate it. At _papi_hwi_shutdown_global_threads() time, who cares, its closing time.