Correct versioning of libpapi.so
The configure for linux always set the soname to libpapi.so.  This
causes problems when /sbin/ldconfig tries to update the library
information on linux.  The shared library is installed as
/lib{64}/libpapi.so.$VERSION, but the shared library has the soname of
libpapi.so.  ldconfig makes a symbolic link from /lib/libpapi.so to
the actual versioned shared library, /lib/{64}/libpapi.so$VERSION.
The configure should get the soname correct to avoid creating this
symbolic link.

This patch only addresses the issues for some of the possible
platforms and similar patches may be needed for other platforms.

Signed-off-by: William Cohen <wcohen@redhat.com>