Add initial support for IBM POWER8 processor
Add initial support for IBM POWER8 processor

The IBM POWER8 processor (to be publicly announced at some
future date) has some preliminary support in libpfm with
a subset of native events.  These POWER8-related libpfm
changes were pulled into PAPI on July 3, so further updates
in PAPI were required to support this new processor.  This
patch adds that required support.  NOTE: Due to the fact
that only a subset of native events have been publicised
at this point (and pushed into libpfm), not all of the
usual PAPI preset events have corresponding native events.
The rest of the POWER8 native events will be pushed
upstream once they are verified, and then we can flesh out
the PAPI preset events.

With this initial POWER8 support patch, 5 of the ctests
and ftests fail, compared to 3 when PAPI is run on a POWER7.
At least one of the failing testcases is due to testing
being done on an early POWER8 processor with some known
hardware problems. We presume the number of failing tests
will decrease once we have GA-level hardware to test on.

Signed-off-by: Maynard Johnson <maynardj@us.ibm.com>