Trial fix for the cuda component static libc linking issue.
Weak link against _dl_non_dynamic_init, this appears in my limited testing to be
in gnu libc.a and not in the so.

For background, it was reported by Steve Kaufmann that statically linking
tools with a PAPI library configured with the CUDA component segfaulted.
It appears that calling any of the dynamic linker functions from a static
executable is asking for pain.

See Trac bug 182 https://icl.cs.utk.edu/trac/papi/ticket/182