add os_detect() callback to pfmlib_pmu struct
Add new callbacks to invoke an OS-specific callback
to detect the PMU. Even though a PMU can be detected
via a CPUID that does not mean that it is also supported
by the host kernel. So proivide an opportunity for PMU
support to invoke an OS specific detection routine.

For instance, on Linux with perf_events, a PMU is
supported by the kernel if it has an entry in /sys/devices/

Signed-off-by: Stephane Eranian <eranian@gmail.com>