perf_event libpfm4 events -- correctly handle invalid events
It was possible for event names to be obtained from libpfm4 during
enumeration that were not valid events.  This usually happens
with uncore events, where the uncore is listed as available
based on cpuid but when libpfm4 tries to get the uncore type from
the kernel finds out it is unsupported.

This change makes this properly fail, instead of just returning "0"
for all the event paramaters (which is a valid event on x86).

Also make this change in the regular perf_event component, even though
it is less likely to happen in practice.