papi_events.csv: First draft preset events on HSW
Contributed by Nils Smeds
-------------------------
Here is a suggestion for addition to Hsw counters. These are not
rigorously tested. It compiles and loads.
I'm rather uncertain on many of the events so I am hoping that adding
events like this will get some useful
feedback from the community so that we can improve.
-------------------------