fix bug with PFM_OS_PERF_EVENT attributes
There was a bug in the  perf_event_mods[] attribute table
that would cause the fully qualified event string to miss
certain attributes such as mg, mh. This was due to the fact
that in a pfmlib_attr_desc_t table the entries must be ordered
by the values of the respective PERF_ATTR_XX because there is
direct indexing by attribute value.

This patch adds a PERF_ATTR_SKIP pseudo attribute to pad the
entries and make the mg, mh aligned in the table with their
corresponding PERF_ATTR_XX macros.

The perf_get_os_attr_info() and perf_get_os_nattrs() are modified
to skip entries.

Note that PERF_SKIP_ENTRIES is ONLY supported for the
perf_event OS attributes for now.

Signed-off-by: Stephane Eranian <eranian@gmail.com>