papi_events.csv: Add PAPI_L1_ICM for Haswell
Thanks to Maurice Marks of Unisys for the contribution
-------------
I've continued testing on Haswell. By comparison with Vtune and Emon on
Haswell I found that we can use
the counter L2_RQSTS:ALL_CODE_RD for PAPI_L1_ICM, which is a very useful
measure.

Attached is my current version of papi_events.csv with Haswell fixes.
-------------