configure: refactor CTEST_TARGETS
Problem:
The set of ctests to build is determined at configure time, in CTEST_TARGETS.
This is set in each OS detection section and suffers from neglect.

Solution:
Try to push the decisions about which tests to build out of configure, ask for them all.
Idealy the tests will be written in such a way as to fail/skip gracefully if they lack
functionallity, teething problems are expected initially.