add PFMLIB_PMU_FL_NO_SMPL pmu flag
Add PFMLIB_PMU_FL_NO_SMPL as a generic PMU flag
(pmu->flags) to indicate that the PMU does not
support sampling.

This flag can be used to prohibit using sampling
related modifiers in the OS layer, for instance.

The patch updates the SNB-EP uncore PMUs to use this
flag because they do not support sampling.

Signed-off-by: Stephane Eranian <eranian@gmail.com>