add Intel RAPL PMU support
This patch adds support for the Intel RAPL energy consumption
counters present on Intel SandyBridge, IvyBridge and Haswell
processors.

The PMU is detected based on OS-support, i.e., the
/sys/devices/power directory must exist.

The following RAPL events are defined:
- RAPL_ENERGY_CORES
- RAPL_ENERGY_PKG
- RAPL_EENRGY_DRAM (servers only)

The new PMU is called rapl.

The library does not provide any unit conversion for the
RAW count. The kernel exports the raw count as increments
of 1/(2^32) Joules or about 0.23 nano-Joules.

Signed-off-by: Stephane Eranian <eranian@gmail.com>