add preliminary aarch64 (arm64) support
There has been some work to build fedora 19 on 64-bit arm armv8 machines
(aarch64).  I took a look that the why the papi build was failing.  The
attached is a set of minimal patches to get papi to build.  The patch is
just a step toward getting aarch64 support for papi. Things are not all
there for papi to work in that environment.  Still need libpfm to
support aarch64 and papi_events.csv describing mappings to machine
specific events.