Make the testsuite as a stand-alone copy-able directory of code
These changes to the Makefiles allows the testsuite to be compiled
separately from the papi sources.  This is useful for people wanting
to experiment with the tests and verify that the existing installation
of papi works.

We put absolute paths to the installed library and include files
into the installed makefile for the tests.