Check the return values of PAPI_start() and PAPI_stop() for the data_range test
The ia64 data_range test did not check the return values of
PAPI_start() or PAPI_stop().  There are propbably few people running
this test on ia64 machine, but this is more to eliminate a couple
errors noted by a Coverity scan and reduce the clutter in the Coverity
scan.

Signed-off-by: William Cohen <wcohen@redhat.com>