Make data_range test use prginfo
Coverity complained about prginfo being an unused variable for
data_range.c.  The code is modified to be stylistically like the code
for hw_info in the preceding lines which also is not used elsewhere in
the test.  This is more to reduce the amount of output in the Coverity
scan than to fix this minor issue.

Signed-off-by: William Cohen <wcohen@redhat.com>