Pass the actual size of the buffer to perf_read_buffer()
A Coverity scan pointed out that the argument for the sizeof() for a
perf_read_buffer() call was suspect.  The perf_read_buffer() third
argument is the size of the buffer pointed to in the second argument.
Other places in the code are doing that and this call should also.

Signed-off-by: Will Cohen <wcohen@redhat.com>