add LIBPFM_DISABLED_PMUS env variable to blacklist PMU models
This patch adds a new environment variable to libpfm.
It allows users or sysadmins to disable certain PMU
models even though they are compiled in. This may be
useful for testing or if the PMU is already supported
via another interface.

The variable takes a comma-separated list of PMU model
names (as they appear in showevtinfo). For instance:

$ LIBPFM_DISABLED_PMUS=core,snb showevtinfo

Signed-off-by: Stephane Eranian <eranian@gmail.com>