PAPI_Matlab: Change outdated API call
Hadl Moghaddam reported an issue with MATLAB support.
It appears that mxCreateScalarDouble has been removed and
mxCreateDoubleScalar is the prefered replacement.

A quick search confirms:
http://www.mathworks.com/matlabcentral/answers/33573
----------------------------------------
mxCreateScalarDouble is an old MEX API that was replaced circa 2002 and
dropped after 2006. You may be able to swap in its replacement, called
mxCreateDoubleScalar, in the file klu_mex.c. You may want to alert the
author of this compatibility problem.

See these old release notes for details on the API change.