add support for ARM Qualcomm Krait PMU
Version 2 of this patch properly detects we have a Krait (not Scorpion)
CPU and also updates the number of counters to reflect what the kernel
reports.

The following patch adds initial support for Qualcomm Krait (Snapdragon)
PMU to libpfm4.

This is based off of the code from the recent
   [PATCH 0/7] Support Krait CPU PMUs
thread on the linux-kernel mailing list.
As far as I can tell there's no publicly available documentation for this
PMU.

We have a PAPI user who is running PAPI on such a processor.

The chip is ARMv7, similar to the Cortex A15.  Some issues:

1. Older versions of the chip have broken PC_WRITE event support.

   Should libpfm4 detect this (by part number) and have separate event
   tables?

2. This patch only has minimal event support for common armv7 events.
   The Krait CPUs support more events through a more complex
   custom encoding.  I think we should wait until the patches for
   this make it into upstream linux-kernel before we bother
   implementing them.

Signed-off-by: Vince Weaver <vincent.weaver@maine.edu>