perf_event_uncore: fixes SNBEP UNC event listing
Gary Mohr, looking at Uncore event enumeration.
-----------------------------------------------------
The problem was that in the SNBEP uncore some masks require other attributes to
be provided. For instance the NID mask requires a nf=x attribute to be provided.
When the attribute was not provided it caused the libpfm4 encode function to
return an error which caused PAPI to stoplooking for umasks in that event.

The biggest change in this area is that when enumerating event masks, the masks
were being used one at a time with the event and the libpfm4 code to encode the
event was being called.

This patch adds code to detect when additional attributes are required and adds
them to the event string before calling the libpfm4 encode function.
-----------------------------------------------------

Many thanks to Gary for diving into this.