perf_event_uncore: More work on Uncore enum
Per Gary Mohr
----------------
The problem listing uncore events which do not have a default umask
declared has been bugging me so I spend some time looking into it today
and found a fix.  It turns out that the error check that I commented out
in libpfm4 to get this to work exists in both the perf_events and the
perf_events_uncore versions of libpfm4.  So I went looking to see if the
core events always declared a default mask and found they do not.

So after looking around a while, I discovered that the papi components
perf_event and perf_event_uncore use this libpfm4 interface differently.
The error is returned from the pfm_get_os_event_encoding libpfm4
function.

The papi/perf_event code follows the call with this test:
if (ret==PFM_SUCCESS) {

The papi/perf_event_uncore code follows the call with this test:

 /* If we error out on UMASK then enumeration doesn't work */
 if ((ret==PFM_SUCCESS) || (ret==PFM_ERR_UMASK)) {
--------------------