Add x87 counts to FP_INS and FP_OPS on [S|I]VB
In Sandy/Ivy Bridge processors it is safe to assume 3 general counters
and the definition of FP_OPS/INS was only using 2. This commit changes
 the definition PAPI_FP_INS/OPS to include FP_COMP_OPS_EXEC:X87

The effect appears minimal and improves counts with naively compiled
LAPACK. ( gfortran version 4.6 on an IvyBridge with the default build
paramaters for LAPACK produced no SSE/AVX ins, it did all its work
with X87 ins)

If issues arise, this is safe to revert.