native_avail: Add a --validate flag
--validate attempts to decide if each event can be added to an eventset.

Patch by Steve Kaufmann
-----------------------
I have modified papi_native_avail to accept a -v (validate) option that
when set will mark an event as actually available or not if it is
attempted to be added to an event set. I did this rather quickly but it
conveys an idea that would inform users that, yes this event name
exists but if you try to access it on the hardware that you are running
papi_native_avail on, you will not get it. For whatever reason. If you
think it a useful feature perhaps consider it for inclusion in papi.