activate perf_event_ext cpu= modifier
Enable cpu=X  perf_events modifier on events.
The cpu number is returned in the pfm_perf_encode_arg.cpu
field. If modifier not set in event string, then field has
value -1.

This is an indicative information of where the user
wants the event to be programmed. This is useful
only in system-wide mode. The library does not
program the event with perf_events. The
user program still has to invoke the perf_event_open()
syscall.

Signed-off-by: Stephane Eranian <eranian@gmail.com>