x86_cpuid_info.c:
Thanks to Gleb Smirnoff of the freebsd project
for pointing out issues with our cpuid implementation on freebsd.
-----------------------------------------------------------------
(speaking about the implementation of cpuid2() )

However, this implementation can't be guaranteed to work on amd64,
because amd64 ABI allows a 128 byte "red zone" - a region in memory
below %rsp, that can be used by a leaf function. And the clang
compiler on FreeBSD does generate a code that utilizes red zones.
So, your push/pop operation clobbers the first word in the red zone.
Probably gcc on Linux doesn't generate red zone code, that's why
you didn't hit the problem.