papi.c: Debuging improvements.
Thanks to Gary Mohr
-------------------
The attached patch file inserts APIDBG statements at the beginning of
most (maybe all) of the API functions that would normally be called by a
PAPI application.  I have made these kind of changes on two or three
different times in the past when working on a PAPI enhancement or trying
to fix a problem we were seeing.  They make it much easier to figure out
what PAPI services are being used by the application.  I think they make
the traces easier to read.

This time I decided to share these changes.  Then if you choose to
commit them, I will not have to put them in again next time I need to
figure out what is going on in PAPI.
-------------------