Fix excluded files for run_tests
Gary Mohr pointed out that in a Makefile refactor we neglected to
update the tests eclude creteria
--------------------------------
It turns out that you guys have made changes in the way the tests
are built which have introduced errors when running the scripts.
The script attempted to automatically remove the makefiles from
the list of files it would execute but your build changes broke
the makefiles up into several files and renamed all of them.
So the script was trying to execute them.

I decided the most flexible way to handle this is to remove the
code from the script that looks for makefiles and just add them
to the exclude files used by the script.
The script will not execute any file listed in the exclude file.

The attached patch implements these changes so the script runs
correctly with the current papi build files.

Gary
--------------------------------