native_avail.c: Bug fixes and updates
Thanks to Gary Mohr
--------------------------------------------------
This patch fixes a couple of problems found in the papi_native_avail program.

First change fixes a problem introduced when the -validate option was added.  This
option causes events to get added to an event set but never removes them.  This change
will remove them if the add works.  This change also fixes a coverity detected error
where the return value from PAPI_destroy_eventset was being ignored.

Second change improves the delimitor check when separating the event description from
the event mask description.  The previous check only looked for a colon but some of the
event descriptions contain a colon so descriptions would get displayed incorrectly.  The
new check finds the "masks:" substring which is what papi inserts to separate these two
descriptions.

Third change adds code to allow the user to enter events of the form pmu:::event or
pmu::event when using the -e option in the program.