papi_internal.c: change SUBDBG to INTDBG
Thanks to Gary Mohr
-------------------
This patch contains changes to replace calls to SUBDBG with calls to INTDBG in this
source file.  This source file should be using the Internal debug macro rather than the
Substrate debug macro so that the PAPI debug filters work correctly.  These changes also
add some new debug calls so that we will get a better picture of what is going on in the
PAPI internal layer.  There are a few calls to the SUBDBG macro that are in code that
I have modified to add support for new event level masks which are not converted by this
patch.  They will be corrected when the event level mask patch is provided.